Alterações com a equipe;

+++ src/Application/Interfaces/Repositories/IAccountsRepository.cs
@@ -7,5 +7,5 @@ public interface IAccountsRepository
 	Task<bool> CreateAsync(Account account);
 	Task<decimal> GetAccountBalanceAsync(string clientEmail);
 	Task<bool> DepositAsync(string clientEmail, decimal amount);
-	Task<bool> WithdrawAsync(string clientEmail, int amount);
+	Task<bool> WithdrawAsync(string clientEmail, decimal amount);
 }
